EmailField - Name Support
When EmailField::filterMessage is invoked the EmailField object attempts to look up the submitter's name and use that as the name set in the message's Reply-To.  It does this by scanning the FieldSubmission entities associated with the provided Submission entity.  If it finds a FieldSubmission entity associated with a Field entity with a type of "name" it uses the "name" property of that FieldSubmission entity's value as the submitter's name.  Otherwise it defaults to using no name (i.e. null) which is what it did unconditionally before this commit.