FormInterface::getEmail
Splits the functionality of generating a Swift_Message instance out of FormInterface::submit and also removes responsibility for actual mail submission from FormInterface instances (this means they do not have to maintain a reference to a Swift_Mailer instance internally and therefore means that there's no need to mock one out to test).

This means that a Submission entity may be created recording all data of a submission even when sending the associated email fails.  This is new reflected in SubmitController: A Submission entity is first created and persisted along with all FieldSubmission entities thereby ensuring the data is persisted.  An attempt is then made to send the associated email and if this succeeds an Email entity is created and attached to the Submission entity (which is then re-persisted thereby ensuring this data is persisted also).